Ouch, rejigging reviews and special offers.

+++ administrator/components/com_reviews/views/reviews/view.html.php
@@ -51,22 +51,35 @@ class ReviewsViewReviews extends JViewLegacy {
    */
   protected function addToolBar() {
     $document = JFactory::getDocument();
+		$canDo = ReviewsHelper::getActions();
 
-    JToolBarHelper::addNew('review.add', 'JTOOLBAR_NEW');
-    JToolBarHelper::editList('review.edit', 'JTOOLBAR_EDIT');
-    JToolBarHelper::publish('reviews.publish', 'JTOOLBAR_PUBLISH', true);
-    JToolBarHelper::unpublish('reviews.unpublish', 'JTOOLBAR_UNPUBLISH', true);
-    JToolBarHelper::trash('reviews.trash');
-    JToolBarHelper::deleteList('Are you sure?', 'reviews.delete', 'JTOOLBAR_DELETE');
-    // Set the title which appears on the toolbar 
-    JToolBarHelper::title(JText::_('Manage reviews'));
- 
+    if ($canDo->get('core.create')) {
+      JToolBarHelper::addNew('review.add', 'JTOOLBAR_NEW');
+    }
+    
+    if ($canDo->get('core.edit')) {
+      JToolBarHelper::editList('review.edit', 'JTOOLBAR_EDIT');
+    }
+    
+    if ($canDo->get('core.edit.state')) {
+      JToolBarHelper::publish('reviews.publish', 'JTOOLBAR_PUBLISH', true);
+      JToolBarHelper::unpublish('reviews.unpublish', 'JTOOLBAR_UNPUBLISH', true);
+      JToolBarHelper::trash('reviews.trash');
+    }
+    
+    if ($canDo->get('core.edit.delete')) {
+      JToolBarHelper::deleteList('Are you sure?', 'reviews.delete', 'JTOOLBAR_DELETE');
+    }
+    
+    // Set the title which appears on the toolbar
+    JToolBarHelper::title(JText::_('COM_REVIEW_VIEW_REVIEWS'));
+
+    
     JHtmlSidebar::addFilter(
 			JText::_('JOPTION_SELECT_PUBLISHED'),
 			'filter_published',
 			JHtml::_('select.options', JHtml::_('jgrid.publishedOptions'), 'value', 'text', $this->state->get('filter.published'), true)
-		);
-		    
+		);	    
   }
